Moves polygon clipping functions into own folder
