+++ src/GeometryOps.jl
@@ -36,6 +36,10 @@ include("methods/equals.jl")
 include("methods/geom_in_geom.jl")
 include("methods/geom_on_geom.jl")
 include("methods/orientation.jl")
+include("methods/clipping/clipping_processor.jl")
+include("methods/clipping/intersection.jl")
+include("methods/clipping/difference.jl")
+include("methods/clipping/union.jl")
 
 include("transformations/flip.jl")
 include("transformations/simplify.jl")
