+++ src/methods/intersects.jl
@@ -1,6 +1,6 @@
 # # Intersection checks
 
-export intersects, intersection, intersection_points, union_test, difference_test
+export intersects, intersection, intersection_points, union, difference
 
 #=
 ## What is `intersects` vs `intersection` vs `intersection_points`?
@@ -51,42 +51,6 @@ geometry dispatches are _line_intersects, which determines if two line segments
 intersect and _intersection_point which determines the intersection point
 between two line segments.
 =#
-
-
-# TODO: tempory poly_in_poly, only works if entirely contained
-# Checks if polya in poly b
-function poly_in_poly(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b)
-    p_a = to_edges(poly_a)[1][1]
-    return point_in_polygon(p_a, poly_b)
-end
-
-poly_in_poly(geom_a, geom_b) =
-    poly_in_poly(GI.trait(geom_a), geom_a, GI.trait(geom_b), geom_b)
-
-# TODO: temporary alternative for to_points, repeats first and last point
-function my_get_pts(::GI.PolygonTrait, poly_a)
-    edges = to_edges(poly_a)
-    len =  length(edges) + 1
-    poly_points = Vector{Tuple{Float64, Float64}}(undef, len)
-    for ii in eachindex(edges)
-        poly_points[ii] = edges[ii][1]
-    end
-    poly_points[end] = edges[1][1]
-    return poly_points
-end
-
-my_get_pts(poly_a) = my_get_pts(GI.trait(poly_a), poly_a)
-
-function lin_ring_to_poly(lin_ring)
-    edges = to_edges(lin_ring)
-    points = Vector{Tuple{Float64, Float64}}(undef, length(edges)+1)
-    for ii in eachindex(edges)
-        points[ii] = edges[ii][1]
-    end
-    points[end] = edges[1][1]
-    return GI.Polygon([points])
-end
-
 """
     intersects(geom1, geom2)::Bool
 
@@ -234,794 +198,6 @@ intersection(
     geom_b,
 ) = intersection_points(trait_a, geom_a, trait_b, geom_b)
 
-function build_a_list(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b, edges_a, edges_b,
-                      intr_list, a_idx_list, b_idx_list, alpha_a_list, alpha_b_list, a_list, b_list)
-    # Find intersection points and adds them to a_list
-    # "counter" is used to index all inter-related lists
-    counter = 1
-    # "acount" is used to index a_list
-    acount = 1
-    for ii in eachindex(edges_a)
-        # Add the first point of the edge to the list of points in a_list
-        if acount <= length(a_list)
-            a_list[acount] = PolyNode(ii, false, 0, false, 0, false)
-        else
-            push!(a_list, PolyNode(ii, false, 0, false, 0, false))
-        end
-        acount = acount + 1
-
-        # Keep track of current position in inter-related lists
-        # before finding new intersection points on our new edge
-        prev_counter = counter
-
-        for jj in eachindex(edges_b)
-
-            # Add the first point of the edge to b_list
-            if ii == 1 
-                b_list[jj] = PolyNode(jj, false, 0, false, 0, false)
-            end
-
-            # Check if edge jj of poly_b intersects with edge ii of poly_a
-            if _line_intersects([edges_a[ii]], [edges_b[jj]]);
-                
-                int_pt, alphas = _intersection_point(edges_a[ii], edges_b[jj])
-                # Store the cartesion coordinates of intersection point
-                intr_list[counter] = int_pt
-                # Store which edge of poly_b the intersection point lies on
-                b_idx_list[counter] = jj
-                # Store the alpha values
-                alpha_a_list[counter] = alphas[1]
-                alpha_b_list[counter] = alphas[2]
-
-                counter = counter + 1
-            else
-                continue
-            end
-        end
-
-        # After iterating through all edges of poly_b for edge ii of poly_a,
-        # add the intersection points to a_list in CORRECT ORDER if we found any
-        if prev_counter < counter
-            # Order intersection points based on alpha values
-            new_order = sortperm(alpha_a_list[prev_counter:counter-1])
-            pts_to_add = Array{PolyNode, 1}(undef, counter - prev_counter)
-            for kk in eachindex(new_order)
-                # Create PolyNodes of the new intersection points in the correct order
-                # and store the correct index in a_idx_list
-                pts_to_add[new_order[kk]] = PolyNode(prev_counter+kk-1, true, 0, false, alpha_a_list[prev_counter+kk-1], false)
-                a_idx_list[prev_counter+kk-1] = acount + new_order[kk] - 1;
-            end
-
-            # Add the PolyNodes to a_list and update acount
-            splice!(a_list, acount:(acount-1), pts_to_add)
-            acount = acount + counter - prev_counter
-            
-        end
-        
-    end
-
-    # Truncate the inter_related lists
-    intr_list = intr_list[1:counter-1] 
-    a_idx_list = a_idx_list[1:counter-1]
-    b_idx_list = b_idx_list[1:counter-1]
-    alpha_a_list = alpha_a_list[1:counter-1]
-    alpha_b_list = alpha_b_list[1:counter-1]
-    return intr_list, a_idx_list, b_idx_list, alpha_a_list, alpha_b_list, counter
-end
-
-function build_b_list(b_list, intr_list, a_idx_list, b_idx_list, alpha_b_list, counter, k)
-    # Iterate through the b_list and add in intersection points
-    # Occasionally I need to skip the new points I added to the array
-    skip = false
-    num_skips = 0
-    b_neighbors = Array{Int, 1}(undef, k)
-    for ii in 1:(length(b_list)+length(intr_list))
-        # TODO: this skipping scheme could be made nicer
-        if skip
-            num_skips = num_skips - 1
-            if num_skips == 0
-                skip = false
-            end
-            continue
-        end
-        # Find the index in the intr_list (same as b_idx_list) where the inter point is
-        # TODO:find all is inefficient though, so it might be better to make dictionary of indices
-        i = findall(x->x==b_list[ii].idx, b_idx_list)
-        if !isempty(i)     
-            # Order intersection points based on alpha values
-            new_order = sortperm(alpha_b_list[i])
-            pts_to_add = Array{PolyNode, 1}(undef, length(i))
-            for m in eachindex(i)
-                pts_to_add[new_order[m]] = PolyNode(i[m], true, a_idx_list[i[m]], false, alpha_b_list[i[m]], false)
-                b_neighbors[i[m]] = ii + new_order[m]
-            end   
-            # I use splice instead of insert so I can insert array   
-            splice!(b_list, ii+1:ii, pts_to_add)
-            skip = true
-            num_skips = length(i)
-        end
-    end
-
-    b_neighbors = b_neighbors[1:counter-1]
-
-    return b_neighbors, b_list
-end
-
-function flag_ent_exit(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b, a_list, b_list, edges_a, edges_b)
-    # Put in ent exit flags for poly_a
-    status = false
-    for ii in eachindex(a_list)
-        if ii == 1
-            temp = point_in_polygon(edges_a[ii][1], poly_b)
-            status = !(temp[1])
-            continue
-        end
-        if a_list[ii].inter
-            a_list[ii].ent_exit = status
-            status = !status
-        end
-    end
-
-    # Put in ent exit flags for poly_b
-    status = false
-    for ii in eachindex(b_list)
-        if ii == 1
-            temp = point_in_polygon(edges_b[ii][1], poly_a)
-            status = !(temp[1])
-            continue
-        end
-        if b_list[ii].inter
-            b_list[ii].ent_exit = status
-            status = !status
-        end
-    end
-
-    return a_list, b_list
-end
-
-
-function build_ab_list(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b)
-    # Make a list for nodes of each polygon. Note the definition of PolyNode
-    a_list = Array{PolyNode, 1}(undef, _nedge(poly_a))
-    b_list = Array{PolyNode, 1}(undef, _nedge(poly_b))
-
-    # Initialize arrays to keep track of the important information 
-    # associated with the intersection points of poly_a and poly_b.
-    # I initialize these arrays assumed there will be a maximum of 
-    # 30 intersection points and then I truncate the arrays later.
-    k = 30
-    # intr_list stores the cartesian coordintes of the intersection point
-    intr_list = Array{Tuple{Real, Real}, 1}(undef, k)
-    # At index i, a_idx_list stores the index of the ith intersection point in a_list
-    a_idx_list = Array{Int, 1}(undef, k)
-    # At index i, b_idx_list stores the edge number of poly_b that the ith intersection
-    # point lies on.
-    b_idx_list = Array{Int, 1}(undef, k)
-    # Alpha values are used to determine the order in which to place
-    # intersection points in a polygon list, which is especially
-    # useful when many intersection points lie on the same edge of a polygon.
-    alpha_a_list = Array{Real, 1}(undef, k)
-    alpha_b_list = Array{Real, 1}(undef, k)
-
-    # These lists store the cartesian coordinates of poly_a and poly_b
-    edges_a = to_edges(poly_a)
-    edges_b = to_edges(poly_b)
-
-    intr_list, a_idx_list, b_idx_list, alpha_a_list, alpha_b_list, counter = build_a_list(GI.trait(poly_a), poly_a, GI.trait(poly_b), poly_b, edges_a, edges_b,
-                                                                                  intr_list, a_idx_list, b_idx_list, alpha_a_list, alpha_b_list, a_list, b_list)
-    b_neighbors, b_list = build_b_list(b_list, intr_list, a_idx_list, b_idx_list, alpha_b_list, counter, k)
-
-    
-    # Iterate through a_list and update the neighbor indices
-    for ii in eachindex(a_list)
-        if a_list[ii].inter
-            a_list[ii].neighbor = b_neighbors[a_list[ii].idx]
-        end
-    end
-
-    # Flag the entry and exists
-    a_list, b_list = flag_ent_exit(GI.trait(poly_a), poly_a, GI.trait(poly_b), poly_b, a_list, b_list, edges_a, edges_b)
-
-    return a_list, b_list, a_idx_list, intr_list, edges_a, edges_b
-end
-
-function trace_intersection(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b, a_list, b_list, a_idx_list, intr_list, edges_a, edges_b)
-    # Pre-allocate array for return polygons
-    return_polys = Vector{Vector{Tuple{Float64, Float64}}}(undef, 0)
-    # TODO: Make sure can remove this counter
-    # counter = 0
-
-    # Keep track of number of processed intersection points
-    processed_pts = 0
-    tracker = copy(a_idx_list)
-
-    while processed_pts < length(intr_list)
-        # Create variables "list_edges" and "list" so that we can toggle between
-        # a_list and b_list
-        list_edges = edges_a
-        list = a_list
-
-        # Find index of first unprocessed intersecting point in subject polygon
-        starting_pt = minimum(tracker)
-        idx = starting_pt
-
-        # Get current first unprocessed intersection point PolyNode
-        current = a_list[idx]
-        # Initialize array to store the intersection polygon cartesian points
-        pt_list = Vector{Tuple{Float64, Float64}}(undef, 0)
-        # Add the first point to the array
-        push!(pt_list, (intr_list[current.idx][1], intr_list[current.idx][2]))
-        
-        # Mark first intersection point as processed
-        processed_pts = processed_pts + 1
-        tracker[current.idx] = typemax(Int)
-
-        current_node_not_starting = true
-        while current_node_not_starting # While the current node isn't the starting one
-            status2 = false
-            current_node_not_intersection = true
-            while current_node_not_intersection # The current node isn't an intersection
-                
-                if current.inter
-                    status2 = current.ent_exit
-                end
-
-                # Depending on status of first intersection point, either
-                # traverse polygon forwards or backwards
-                if status2
-                    idx = idx + 1
-                else
-                    idx = idx -1
-                end
-
-                # Wrap around the point list
-                if idx > length(list)
-                    idx = mod(idx, length(list))
-                elseif idx == 0
-                    idx = length(list)
-                end
-
-                # Get current node
-                current = list[idx]
-
-                # Add current node to the pt_list
-                if current.inter
-                    # Add cartesian coordinates from inter_list
-                    push!(pt_list, (intr_list[current.idx][1], intr_list[current.idx][2]))
-                    
-                    # Keep track of processed intersection points
-                    if (current != a_list[starting_pt] && current != b_list[a_list[starting_pt].neighbor])
-                        processed_pts = processed_pts + 1
-                        tracker[current.idx] = typemax(Int)
-                    end
-                    
-                else
-                    # Add cartesian coordinates from "list", which should point to either a_list or b_list
-                    push!(pt_list, (list_edges[current.idx][1][1], list_edges[current.idx][1][2]))
-                end
-
-                current_node_not_intersection = !current.inter
-            end
-            
-            # Break once get back to starting point
-            current_node_not_starting = (current != a_list[starting_pt] && current != b_list[a_list[starting_pt].neighbor])
-
-            # Switch to neighbor list
-            if list == a_list
-                list = b_list
-                list_edges = edges_b
-            else
-                list = a_list
-                list_edges = edges_a
-            end
-            idx = current.neighbor
-            current = list[idx]
-        end
-
-        push!(return_polys, pt_list)
-        # TODO: Make sure can remove this counter
-        # counter = counter + 1
-    end
-
-    # Check if one polygon totally within other
-    # TODO: use point list instead of edges once to_points is fixed
-    if isempty(return_polys)
-        if point_in_polygon(edges_a[1][1], poly_b)[1]
-            list = []
-            for i in eachindex(edges_a)
-                push!(list, edges_a[i][1])
-            end
-            push!(list, edges_a[1][1])
-            push!(return_polys, list)
-        elseif point_in_polygon(edges_b[1][1], poly_a)[1]
-            list = []
-            for i in eachindex(edges_b)
-                push!(list, edges_b[i][1])
-            end
-            push!(list, edges_b[1][1])
-            push!(return_polys, list)
-        end
-    end
-    return return_polys
-end
-
-function get_inter_holes(return_polys, poly_a, ext_poly_a, poly_b, ext_poly_b)
-    final_polys =  Vector{Vector{Vector{Tuple{Float64, Float64}}}}(undef, 0)
-
-    for poly in return_polys
-        # Turning poly into a Vec Vec Vec so i can put more polys in it
-        poly = [[poly]]
-        for hole in GI.gethole(poly_a) 
-            replacement_p = Vector{Vector{Vector{Tuple{Float64, Float64}}}}(undef, 0)
-            for p in poly
-                new_ps = difference_test(GI.Polygon(p), lin_ring_to_poly(hole))
-                append!(replacement_p, new_ps)
-            end
-            poly = replacement_p
-        end
-        
-        for hole in GI.gethole(poly_b)
-            replacement_p = Vector{Vector{Vector{Tuple{Float64, Float64}}}}(undef, 0)
-            for p in poly
-                new_ps = difference_test(GI.Polygon(p), lin_ring_to_poly(hole))
-                append!(replacement_p, new_ps)
-            end
-            poly = replacement_p
-        end
-        
-        append!(final_polys, poly)
-    end
-
-    return final_polys
-        
-end
-
-function get_union_holes(return_polys, poly_a, ext_poly_a, poly_b, ext_poly_b)
-    # This will be of length one, but I'm keeping it this type to be consistent: vec of vec of vec of tup
-    final_polys =  Vector{Vector{Vector{Tuple{Float64, Float64}}}}(undef, 1)
-    
-    # don't need for-loop because union means return_polys only represents 1 poly
-    # if my code is correct, the first poly in return_polys represents the exterior
-    poly = return_polys[1]
-    mult_poly = return_polys
-    
-    for hole in GI.gethole(poly_a)
-        new_hole = difference_test(lin_ring_to_poly(hole), ext_poly_b)
-        for h in new_hole
-            if length(h)>0
-                # I claim I can index h at one, because it can't
-                # have a hole because a hole within a hole would
-                # be an invalid polygon
-                append!(mult_poly, [h[1]])
-            end
-        end        
-    end
-    
-    for hole in GI.gethole(poly_b)
-        new_hole = difference_test(lin_ring_to_poly(hole), poly_a)
-        for h in new_hole
-            if length(h)>0
-                # if !([h[1]] in mult_poly)
-                #     append!(mult_poly, [h[1]])
-                # end
-                append!(mult_poly, [h[1]])
-            end
-        end 
-    end
-    final_polys[1] = mult_poly 
-
-    return final_polys
-        
-end
-
-function get_difference_holes(return_polys, poly_a, ext_poly_a, poly_b, ext_poly_b, diff_polygons)
-    
-    final_polys =  Vector{Vector{Vector{Tuple{Float64, Float64}}}}(undef, 0)
-
-    if !diff_polygons
-        poly = [return_polys]
-        
-        for hole in GI.gethole(poly_a)
-            replacement_p = Vector{Vector{Vector{Tuple{Float64, Float64}}}}(undef, 0)
-            for p in poly
-                new_ps = difference_test(GI.Polygon(p), lin_ring_to_poly(hole))
-                append!(replacement_p, new_ps)
-            end
-            poly = replacement_p
-        end
-        
-        append!(final_polys, poly)
-
-        for hole in GI.gethole(poly_b)
-            append!(final_polys, intersection(lin_ring_to_poly(hole), poly_a))
-        end
-
-    else
-        for poly in return_polys
-            poly = [[poly]]
-            for hole in GI.gethole(poly_a)
-                replacement_p = Vector{Vector{Vector{Tuple{Float64, Float64}}}}(undef, 0)
-                for p in poly
-                    new_ps = difference_test(GI.Polygon(p), lin_ring_to_poly(hole))
-                    append!(replacement_p, new_ps)
-                end
-                poly = replacement_p
-            end
-            
-            append!(final_polys, poly)
-        end
-
-        for hole in GI.gethole(poly_b)
-            append!(final_polys, intersection(lin_ring_to_poly(hole), poly_a))
-        end
-    end
-
-    return final_polys
-        
-end
-
-"""
-    intersection(
-        ::GI.PolygonTrait, poly_a,
-        ::GI.PolygonTrait, poly_b,
-    )::Union{
-        ::Vector{Vector{Tuple{::Real, ::Real}}}, # is this a good return type?
-        ::Nothing
-    }
-
-Calculates the intersection between two line segments. Return nothing if
-there isn't one.
-## Example
-
-```jldoctest
-import GeoInterface as GI, GeometryOps as GO
-
-p1 = GI.Polygon([[(0.0, 0.0), (5.0, 5.0), (10.0, 0.0), (5.0, -5.0), (0.0, 0.0)]])
-p2 = GI.Polygon([[(3.0, 0.0), (8.0, 5.0), (13.0, 0.0), (8.0, -5.0), (3.0, 0.0)]])
-GO.intersection(p1, p2)
-
-# output
-1-element Vector{Vector{Tuple{Float64, Float64}}}:
-[(6.5, 3.5), (10.0, 0.0), (6.5, -3.5), (3.0, 0.0), (6.5, 3.5)]
-```
-"""
-function intersection(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b)
-    ext_poly_a = GI.getexterior(poly_a)
-    ext_poly_a = lin_ring_to_poly(ext_poly_a)
-    ext_poly_b = GI.getexterior(poly_b)
-    ext_poly_b = lin_ring_to_poly(ext_poly_b)
-    a_list, b_list, a_idx_list, intr_list, edges_a, edges_b = build_ab_list(GI.trait(ext_poly_a), ext_poly_a, GI.trait(ext_poly_b), ext_poly_b)
-    polys = trace_intersection(GI.trait(ext_poly_a), ext_poly_a, GI.trait(ext_poly_b), ext_poly_b, a_list, b_list, a_idx_list, intr_list, edges_a, edges_b)
-
-    if GI.nhole(poly_a)==0 && GI.nhole(poly_b)==0
-        return [polys]
-    else
-        return get_inter_holes(polys, poly_a, ext_poly_a, poly_b, ext_poly_b)
-    end    
-
-end
-
-function trace_union(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b, a_list, b_list, a_idx_list, intr_list, edges_a, edges_b)
-    # Pre-allocate array for return polygons
-    return_polys = Vector{Vector{Tuple{Float64, Float64}}}(undef, 0)
-    # TODO: Make sure can remove this counter
-    # counter = 0
-
-    # Keep track of number of processed intersection points
-    processed_pts = 0
-    tracker = copy(a_idx_list)
-
-    while processed_pts < length(intr_list)
-        # Create variables "list_edges" and "list" so that we can toggle between
-        # a_list and b_list
-        list_edges = edges_a
-        list = a_list
-
-        # Find index of first unprocessed intersecting point in subject polygon
-        starting_pt = minimum(tracker)
-        idx = starting_pt
-
-        # Get current first unprocessed intersection point PolyNode
-        current = a_list[idx]
-        # Initialize array to store the intersection polygon cartesian points
-        pt_list = Vector{Tuple{Float64, Float64}}(undef, 0)
-        # Add the first point to the array
-        push!(pt_list, (intr_list[current.idx][1], intr_list[current.idx][2]))
-        
-        # Mark first intersection point as processed
-        processed_pts = processed_pts + 1
-        tracker[current.idx] = typemax(Int)
-
-        current_node_not_starting = true
-        while current_node_not_starting # While the current node isn't the starting one
-            status2 = false
-            current_node_not_intersection = true
-            while current_node_not_intersection # The current node isn't an intersection
-                
-                if current.inter
-                    status2 = current.ent_exit
-                end
-
-                # Depending on status of first intersection point, either
-                # traverse polygon forwards or backwards
-                if status2
-                    idx = idx - 1
-                else
-                    idx = idx + 1
-                end
-
-                # Wrap around the point list
-                if idx > length(list)
-                    idx = mod(idx, length(list))
-                elseif idx == 0
-                    idx = length(list)
-                end
-
-                # Get current node
-                current = list[idx]
-
-                # Add current node to the pt_list
-                if current.inter
-                    # Add cartesian coordinates from inter_list
-                    push!(pt_list, (intr_list[current.idx][1], intr_list[current.idx][2]))
-                    
-                    # Keep track of processed intersection points
-                    if (current != a_list[starting_pt] && current != b_list[a_list[starting_pt].neighbor])
-                        processed_pts = processed_pts + 1
-                        tracker[current.idx] = typemax(Int)
-                    end
-                    
-                else
-                    # Add cartesian coordinates from "list", which should point to either a_list or b_list
-                    push!(pt_list, (list_edges[current.idx][1][1], list_edges[current.idx][1][2]))
-                end
-
-                current_node_not_intersection = !current.inter
-            end
-            
-            # Break once get back to starting point
-            current_node_not_starting = (current != a_list[starting_pt] && current != b_list[a_list[starting_pt].neighbor])
-
-            # Switch to neighbor list
-            if list == a_list
-                list = b_list
-                list_edges = edges_b
-            else
-                list = a_list
-                list_edges = edges_a
-            end
-            idx = current.neighbor
-            current = list[idx]
-        end
-
-        push!(return_polys, pt_list)
-        # TODO: Make sure can remove this counter
-        # counter = counter + 1
-    end
-
-    # Check if one polygon totally within other
-    # TODO: use point list instead of edges once to_points is fixed
-    if isempty(return_polys)
-        if point_in_polygon(edges_a[1][1], poly_b)[1]
-            list = []
-            for i in eachindex(edges_b)
-                push!(list, edges_b[i][1])
-            end
-            push!(list, edges_b[1][1])
-            push!(return_polys, list)
-            return return_polys
-        elseif point_in_polygon(edges_b[1][1], poly_a)[1]
-            list = []
-            for i in eachindex(edges_a)
-                push!(list, edges_a[i][1])
-            end
-            push!(list, edges_a[1][1])
-            push!(return_polys, list)
-            return return_polys
-        else
-            push!(return_polys, my_get_pts(poly_a))
-            push!(return_polys, my_get_pts(poly_b))
-            return return_polys
-        end
-    end
-
-    if length(return_polys) > 1
-        outer_idx = 1
-        for j = 2:(length(return_polys)-1)
-            poly1 = GI.Polygon([return_polys[outer_idx]])
-            poly2 = GI.Polygon([return_polys[j]])
-            if !poly_in_poly(poly2, poly1)
-                outer_idx = j
-            end
-        end
-
-        if outer_idx == 1
-            return return_polys
-        elseif outer_idx == length(return_polys)
-            arr = outer_idx
-            splice!(arr, 2:1, 1:(length(return_polys)-1))
-            return return_polys[arr]
-        else
-            arr = 1:(outer_idx-1)
-            splice!(arr, (length(arr)+1):length(arr), (outer_idx+1):length(return_polys))
-            return return_polys[arr]
-        end
-    end
-
-    return return_polys
-end
-
-function union_test(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b)
-    ext_poly_a = GI.getexterior(poly_a)
-    ext_poly_a = lin_ring_to_poly(ext_poly_a)
-    ext_poly_b = GI.getexterior(poly_b)
-    ext_poly_b = lin_ring_to_poly(ext_poly_b)
-    a_list, b_list, a_idx_list, intr_list, edges_a, edges_b = build_ab_list(GI.trait(ext_poly_a), ext_poly_a, GI.trait(ext_poly_b), ext_poly_b)
-    polys = trace_union(GI.trait(ext_poly_a), ext_poly_a, GI.trait(ext_poly_b), ext_poly_b, a_list, b_list, a_idx_list, intr_list, edges_a, edges_b)
-    if GI.nhole(poly_a)==0 && GI.nhole(poly_b)==0
-        return [polys]
-    else
-        return get_union_holes(polys, poly_a, ext_poly_a, poly_b, ext_poly_b)
-    end 
-end
-
-union_test(geom_a, geom_b) =
-    union_test(GI.trait(geom_a), geom_a, GI.trait(geom_b), geom_b)
-
-function trace_difference(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b, a_list, b_list, a_idx_list, intr_list, edges_a, edges_b)
-    # Pre-allocate array for return polygons
-    return_polys = Vector{Vector{Tuple{Float64, Float64}}}(undef, 0)
-    # TODO: Make sure can remove this counter
-    # counter = 0
-
-    # Keep track of number of processed intersection points
-    processed_pts = 0
-    tracker = copy(a_idx_list)
-
-    while processed_pts < length(intr_list)
-        # Create variables "list_edges" and "list" so that we can toggle between
-        # a_list and b_list
-        list_edges = edges_a
-        list = a_list
-
-        # Find index of first unprocessed intersecting point in subject polygon
-        starting_pt = minimum(tracker)
-        idx = starting_pt
-
-        # Get current first unprocessed intersection point PolyNode
-        current = a_list[idx]
-        # Initialize array to store the intersection polygon cartesian points
-        pt_list = Vector{Tuple{Float64, Float64}}(undef, 0)
-        # Add the first point to the array
-        push!(pt_list, (intr_list[current.idx][1], intr_list[current.idx][2]))
-        
-        # Mark first intersection point as processed
-        processed_pts = processed_pts + 1
-        tracker[current.idx] = typemax(Int)
-
-        current_node_not_starting = true
-        while current_node_not_starting # While the current node isn't the starting one
-            status2 = false
-            current_node_not_intersection = true
-            while current_node_not_intersection # The current node isn't an intersection
-                
-                if current.inter
-                    status2 = current.ent_exit
-                end
-
-                # Depending on status of first intersection point, either
-                # traverse polygon forwards or backwards
-                if (!status2 && list == a_list) || (status2 && list == b_list)
-                    idx = idx + 1
-                else
-                    idx = idx - 1
-                end
-
-                # Wrap around the point list
-                if idx > length(list)
-                    idx = mod(idx, length(list))
-                elseif idx == 0
-                    idx = length(list)
-                end
-
-                # Get current node
-                current = list[idx]
-
-                # Add current node to the pt_list
-                if current.inter
-                    # Add cartesian coordinates from inter_list
-                    push!(pt_list, (intr_list[current.idx][1], intr_list[current.idx][2]))
-                    
-                    # Keep track of processed intersection points
-                    if (current != a_list[starting_pt] && current != b_list[a_list[starting_pt].neighbor])
-                        processed_pts = processed_pts + 1
-                        tracker[current.idx] = typemax(Int)
-                    end
-                    
-                else
-                    # Add cartesian coordinates from "list", which should point to either a_list or b_list
-                    push!(pt_list, (list_edges[current.idx][1][1], list_edges[current.idx][1][2]))
-                end
-
-                current_node_not_intersection = !current.inter
-            end
-            
-            # Break once get back to starting point
-            current_node_not_starting = (current != a_list[starting_pt] && current != b_list[a_list[starting_pt].neighbor])
-
-            # Switch to neighbor list
-            if list == a_list
-                list = b_list
-                list_edges = edges_b
-            else
-                list = a_list
-                list_edges = edges_a
-            end
-            idx = current.neighbor
-            current = list[idx]
-        end
-
-        push!(return_polys, pt_list)
-        # TODO: Make sure can remove this counter
-        # counter = counter + 1
-    end
-
-    diff_polygons = length(return_polys) > 1
-    # # Check if one polygon totally within other
-    # # TODO: use point list instead of edges once to_points is fixed
-    if isempty(return_polys)
-        list_b = []
-        for i in eachindex(edges_b)
-            push!(list_b, edges_b[i][1])
-        end
-        push!(list_b, edges_b[1][1])
-
-        list_a = []
-        for i in eachindex(edges_a)
-            push!(list_a, edges_a[i][1])
-        end
-        push!(list_a, edges_a[1][1])
-
-        if point_in_polygon(edges_a[1][1], poly_b)[1]
-            return return_polys, diff_polygons
-        elseif point_in_polygon(edges_b[1][1], poly_a)[1]
-            push!(return_polys, list_a)
-            push!(return_polys, list_b)
-            return return_polys, diff_polygons
-
-        else
-            # equivalent of push!(return_polys, list_a) since return_polys empty
-            return [list_a], diff_polygons
-        end
-    end
-    return return_polys, diff_polygons
-end
-    
-function difference_test(::GI.PolygonTrait, poly_a, ::GI.PolygonTrait, poly_b)
-    ext_poly_a = GI.getexterior(poly_a)
-    ext_poly_a = lin_ring_to_poly(ext_poly_a)
-    ext_poly_b = GI.getexterior(poly_b)
-    ext_poly_b = lin_ring_to_poly(ext_poly_b)
-    a_list, b_list, a_idx_list, intr_list, edges_a, edges_b = build_ab_list(GI.trait(ext_poly_a), ext_poly_a, GI.trait(ext_poly_b), ext_poly_b)
-    test = trace_difference(GI.trait(ext_poly_a), ext_poly_a, GI.trait(ext_poly_b), ext_poly_b, a_list, b_list, a_idx_list, intr_list, edges_a, edges_b)
-    polys = test[1]
-    diff_polygons = test[2]
-    if GI.nhole(poly_a)==0 && GI.nhole(poly_b)==0
-        final_polys =  Vector{Vector{Vector{Tuple{Float64, Float64}}}}(undef, length(polys))
-        for i in 1:length(polys)
-            final_polys[i] = [polys[i]]
-        end
-        return final_polys
-    else
-        return get_difference_holes(polys, poly_a, ext_poly_a, poly_b, ext_poly_b, diff_polygons)
-    end 
-end
-
-
-
-difference_test(geom_a, geom_b) =
-    difference_test(GI.trait(geom_a), geom_a, GI.trait(geom_b), geom_b)
-
 """
     intersection(
         ::GI.AbstractTrait, geom_a,
@@ -1151,14 +327,4 @@ function _intersection_point((a1, a2)::Tuple, (b1, b2)::Tuple)
         return (x, y), (t, u)
     end
     return nothing, nothing
-end
-
-
-mutable struct PolyNode
-    idx::Int
-    inter::Bool
-    neighbor::Int
-    ent_exit::Bool
-    alpha::Real
-    proc::Bool
 end
